correctly call interaction endpoint; arrays in resolveEndpoint
