+++ applications/fishing-map/src/features/map-features/map-features.hooks.ts
@@ -1,18 +1,23 @@
 import { useSelector, useDispatch } from 'react-redux'
 import GFWAPI from '@globalfishingwatch/api-client'
 import { ExtendedFeature, InteractionEvent } from '@globalfishingwatch/react-hooks'
-import { resolveEndpoint } from '@globalfishingwatch/dataviews-client'
+import { resolveEndpoint, Dataview } from '@globalfishingwatch/dataviews-client'
+import { Generators } from '@globalfishingwatch/layer-composer'
 import {
-  selectTemporalgridDataviews,
   selectWorkspaceDataviews,
-} from '../workspace/workspace.selectors'
+  selectTemporalgridDataviews,
+} from 'features/workspace/workspace.selectors'
+import { selectTimerange } from 'routes/routes.selectors'
 import { setClickedEvent, selectClickedEvent } from './map-features.slice'
 
 export const useClickedEventConnect = () => {
   const dispatch = useDispatch()
   const clickedEvent = useSelector(selectClickedEvent)
+
   const dataviews = useSelector(selectWorkspaceDataviews)
   const temporalgridDataviews = useSelector(selectTemporalgridDataviews)
+  const { start, end } = useSelector(selectTimerange)
+
   const dispatchClickedEvent = (event: InteractionEvent | null) => {
     if (!event || !event.features) return
     // TODO should work for multiple features
@@ -26,6 +31,7 @@ export const useClickedEventConnect = () => {
     const dataset = dataview.datasets?.find((dataset) => dataset.id === DATASET_ID)
     if (!dataset) return []
     const datasetConfig = {
+      endpoint: '4wings-interaction',
       datasetId: DATASET_ID,
       params: [
         { id: 'z', value: feature.tile.z },
@@ -33,8 +39,14 @@ export const useClickedEventConnect = () => {
         { id: 'y', value: feature.tile.y },
         { id: 'rows', value: feature.temporalgrid.row },
         { id: 'cols', value: feature.temporalgrid.col },
+        { id: 'cols', value: feature.temporalgrid.col },
       ],
-      endpoint: '4wings-interaction',
+      query: [
+        // TODO remove hardcoded dataset ID
+        { id: 'datasets', value: ['fishing_v4'] },
+        { id: 'date-range', value: [start, end].join(',') },
+        // { id: 'limit', value: 11 },
+      ],
     }
     const url = resolveEndpoint(dataset, datasetConfig)
     if (url) {
@@ -46,4 +58,49 @@ export const useClickedEventConnect = () => {
     dispatch(setClickedEvent(event))
   }
   return { clickedEvent, dispatchClickedEvent }
+}
+
+export type TooltipEventFeature = {
+  title: string
+  color?: string
+  unit?: string
+  value: string
+}
+
+export type TooltipEvent = {
+  latitude: number
+  longitude: number
+  features: TooltipEventFeature[]
+}
+
+export const useMapTooltip = (event?: InteractionEvent | null) => {
+  const dataviews = useSelector(selectWorkspaceDataviews)
+  const temporalgridDataviews = useSelector(selectTemporalgridDataviews)
+  if (!event || !event.features || !dataviews) return null
+  const tooltipEventFeatures: TooltipEventFeature[] = event.features.flatMap((feature) => {
+    let dataview
+    if (feature.generator === Generators.Type.HeatmapAnimated) {
+      if (!feature.temporalgrid || feature.temporalgrid.sublayerIndex === undefined) {
+        return []
+      }
+
+      // TODO We assume here that temporalgrid dataviews appear in the same order as sublayers are set in the generator, ie indices will match feature.temporalgrid.sublayerIndex
+      dataview = temporalgridDataviews[feature.temporalgrid?.sublayerIndex]
+    } else {
+      dataview = dataviews.find((dataview: Dataview) => dataview.id === feature.generatorId)
+    }
+    if (!dataview) return []
+    return {
+      title: dataview.name || dataview.id.toString(),
+      color: dataview.config.color || 'black',
+      // unit: dataview.unit || '',
+      value: feature.value,
+    }
+  })
+  if (!tooltipEventFeatures.length) return null
+  return {
+    latitude: event.latitude,
+    longitude: event.longitude,
+    features: tooltipEventFeatures,
+  }
 }
